+++ src/scan.cc
@@ -66,6 +66,7 @@ auto search_getwork(uint64_t * countref, uint64_t * firstref) -> bool;
 void search_worker_core(int64_t t);
 auto search_worker(void * vp) -> void *;
 
+
 void search_alloc(struct search_data * sdp)
 {
   constexpr unsigned int one_kilobyte {1024};
@@ -91,6 +92,7 @@ void search_alloc(struct search_data * sdp)
   memset(sdp->dir_array, 0, dirbufferbytes);
 }
 
+
 void search_free(struct search_data * sdp)
 {
   xfree(sdp->qtable);
@@ -100,6 +102,7 @@ void search_free(struct search_data * sdp)
   xfree(sdp->hearray);
   xfree(sdp->dir_array);
 }
+
 
 void search_init(struct search_data * sdp)
 {
@@ -115,6 +118,7 @@ void search_init(struct search_data * sdp)
     sdp->qtable_w[i] = sdp->dprofile_w + word_offset;
   }
 }
+
 
 void search_chunk(struct search_data * sdp, int64_t bits)
 {
@@ -192,6 +196,7 @@ void search_chunk(struct search_data * sdp, int64_t bits)
  }
 }
 
+
 auto search_getwork(uint64_t * countref, uint64_t * firstref) -> bool
 {
   // * countref = how many sequences to search
@@ -219,6 +224,7 @@ auto search_getwork(uint64_t * countref, uint64_t * firstref) -> bool
   return status;
 }
 
+
 #if 0
 
 /* never used */
@@ -236,6 +242,7 @@ void master_dump()
 
 #endif
 
+
 void search_worker_core(int64_t t)
 {
   search_init(sd+t);
@@ -243,6 +250,7 @@ void search_worker_core(int64_t t)
     search_chunk(sd+t, master_bits);
   }
 }
+
 
 void search_do(uint64_t query_no,
                uint64_t listlength,
@@ -294,6 +302,7 @@ void search_do(uint64_t query_no,
   }
 }
 
+
 void search_begin()
 {
   longestdbsequence = db_getlongestsequence();
@@ -309,9 +318,9 @@ void search_begin()
 
   /* start threads */
 
-  search_threads
-    = new ThreadRunner(static_cast<int>(opt_threads), search_worker_core);
+  search_threads = new ThreadRunner(opt_threads, search_worker_core);
 }
+
 
 void search_end()
 {
